Add "-gif" suffix to llama emojis that are .gif
This future proofs the llama emoji set since we might have a .png
version and a .gif version of the same expression:

ie. llama-bored.png and llama-bored-gif.gif

As far as Slack is concerned, these two emoji will be uploaded with the
same name, which will cause a naming conflict on the latter.